+++ src/main/java/com/telepathicgrunt/structurefloaters/StructureFloaters.java
@@ -5,16 +5,17 @@ import com.telepathicgrunt.structurefloaters.mixin.worldgen.StructurePieceAccess
 import me.shedaniel.autoconfig.AutoConfig;
 import me.shedaniel.autoconfig.serializer.JanksonConfigSerializer;
 import net.fabricmc.api.ModInitializer;
+import net.minecraft.structure.PoolStructurePiece;
 import net.minecraft.structure.StructurePiece;
 import net.minecraft.structure.StructureStart;
+import net.minecraft.structure.pool.StructurePool;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.ChunkPos;
+import net.minecraft.util.registry.DynamicRegistryManager;
 import net.minecraft.util.registry.Registry;
 import net.minecraft.world.StructureWorldAccess;
 import net.minecraft.world.gen.chunk.ChunkGenerator;
-import net.minecraft.world.gen.feature.FeatureConfig;
-import net.minecraft.world.gen.feature.OceanMonumentFeature;
+import net.minecraft.world.gen.feature.StructureFeature;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -38,7 +39,7 @@ public class StructureFloaters implements ModInitializer {
 		//Set up config
 		AutoConfig.register(SFConfig.class, JanksonConfigSerializer::new);
 		SF_CONFIG = AutoConfig.getConfigHolder(SFConfig.class).getConfig();
-		STRUCTURES_TO_IGNORE = Arrays.stream(SF_CONFIG.structureToIgnore
+		STRUCTURES_TO_IGNORE = Arrays.stream(SF_CONFIG.configuredStructures
 						.toLowerCase(Locale.ROOT)
 						.replace(" ", "")
 						.replace("	", "")
@@ -46,7 +47,7 @@ public class StructureFloaters implements ModInitializer {
 				.map(Identifier::new)
 				.collect(Collectors.toSet());
 
-		STRUCTURES_TO_RAISE_PIECES_INDIVIDUALLY = Arrays.stream(SF_CONFIG.structureToRaiseEachPieceSeparately
+		STRUCTURES_TO_RAISE_PIECES_INDIVIDUALLY = Arrays.stream(SF_CONFIG.configuredStructuresToRaiseEachPieceSeparately
 						.toLowerCase(Locale.ROOT)
 						.replace(" ", "")
 						.replace("	", "")
@@ -55,12 +56,12 @@ public class StructureFloaters implements ModInitializer {
 				.collect(Collectors.toSet());
 	}
 
-	public static <C extends FeatureConfig> void offsetStructurePieces(StructureStart<C> structureStart, ChunkGenerator generator) {
-		if(structureStart instanceof OceanMonumentFeature.Start) {
+	public static void offsetStructurePieces(DynamicRegistryManager registryManager, StructureStart structureStart, ChunkGenerator generator) {
+		if(structureStart.getFeature().feature == StructureFeature.MONUMENT) {
 			return;
 		}
 
-		Identifier structureID = Registry.STRUCTURE_FEATURE.getId(structureStart.getFeature());
+		Identifier structureID = registryManager.get(Registry.CONFIGURED_STRUCTURE_FEATURE_KEY).getId(structureStart.getFeature());
 		if(STRUCTURES_TO_IGNORE.contains(structureID)) {
 			return;
 		}
@@ -76,8 +77,11 @@ public class StructureFloaters implements ModInitializer {
 			if(y < StructureFloaters.SF_CONFIG.snapStructureToHeight) {
 				structureStart.getChildren().forEach(piece -> {
 					if(raisePiecesSeparately) {
-						if(piece.getBoundingBox().getMinY() <= generator.getMinimumY()) {
+						if(piece.getBoundingBox().getMinY() <= StructureFloaters.SF_CONFIG.snapStructureToHeight) {
 							piece.translate(0, StructureFloaters.SF_CONFIG.snapStructureToHeight - piece.getBoundingBox().getMinY(), 0);
+							if(piece instanceof PoolStructurePiece poolStructurePiece) {
+								poolStructurePiece.getPoolElement().setProjection(StructurePool.Projection.RIGID);
+							}
 						}
 					}
 					else {