+++ src/main/java/com/telepathicgrunt/structurefloaters/mixin/worldgen/OceanMonumentStructureMixin.java
@@ -1,11 +1,6 @@
 package com.telepathicgrunt.structurefloaters.mixin.worldgen;
 
-import net.minecraft.util.math.ChunkPos;
-import net.minecraft.world.HeightLimitView;
-import net.minecraft.world.biome.Biome;
-import net.minecraft.world.biome.source.BiomeSource;
-import net.minecraft.world.gen.ChunkRandom;
-import net.minecraft.world.gen.chunk.ChunkGenerator;
+import net.minecraft.structure.StructureGeneratorFactory;
 import net.minecraft.world.gen.feature.DefaultFeatureConfig;
 import net.minecraft.world.gen.feature.OceanMonumentFeature;
 import org.spongepowered.asm.mixin.Mixin;
@@ -22,16 +17,13 @@ public class OceanMonumentStructureMixin {
      * @reason make Ocean Monuments skip their RIVER/OCEAN category checks if in floating island world. Otherwise, Monuments don't spawn. Mojank lmao
      */
     @Inject(
-            method = "shouldStartAt(Lnet/minecraft/world/gen/chunk/ChunkGenerator;Lnet/minecraft/world/biome/source/BiomeSource;JLnet/minecraft/world/gen/ChunkRandom;Lnet/minecraft/util/math/ChunkPos;Lnet/minecraft/world/biome/Biome;Lnet/minecraft/util/math/ChunkPos;Lnet/minecraft/world/gen/feature/DefaultFeatureConfig;Lnet/minecraft/world/HeightLimitView;)Z",
+            method = "canGenerate(Lnet/minecraft/structure/StructureGeneratorFactory$Context;)Z",
             at = @At(value = "HEAD"),
             cancellable = true
     )
-    private void structurefloaters_modifyBiomeRegistry(ChunkGenerator chunkGenerator, BiomeSource biomeSource, long seed,
-                                                  ChunkRandom chunkRandom, ChunkPos chunkPos, Biome biome,
-                                                  ChunkPos chunkPos2, DefaultFeatureConfig defaultFeatureConfig,
-                                                  HeightLimitView heightLimitView, CallbackInfoReturnable<Boolean> cir)
+    private static void structurefloaters_skipCategoryChecks(StructureGeneratorFactory.Context<DefaultFeatureConfig> context, CallbackInfoReturnable<Boolean> cir)
     {
-        if(chunkGenerator.getSeaLevel() <= chunkGenerator.getMinimumY()){
+        if(context.chunkGenerator().getSeaLevel() <= context.chunkGenerator().getMinimumY()){
             cir.setReturnValue(true);
         }
     }