+++ src/main/java/com/telepathicgrunt/structurefloaters/mixin/worldgen/ShipwreckGeneratorPieceMixin.java
@@ -23,6 +23,7 @@ import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.ModifyVariable;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 import org.spongepowered.asm.mixin.injection.callback.LocalCapture;
 
@@ -50,21 +51,21 @@ public abstract class ShipwreckGeneratorPieceMixin extends SimpleStructurePiece 
      * @reason Prevent structures from being placed at world bottom if disallowed in config
      */
     @Inject(
-            method = "generate(Lnet/minecraft/world/StructureWorldAccess;Lnet/minecraft/world/gen/StructureAccessor;Lnet/minecraft/world/gen/chunk/ChunkGenerator;Ljava/util/Random;Lnet/minecraft/util/math/BlockBox;Lnet/minecraft/util/math/ChunkPos;Lnet/minecraft/util/math/BlockPos;)Z",
+            method = "generate(Lnet/minecraft/world/StructureWorldAccess;Lnet/minecraft/world/gen/StructureAccessor;Lnet/minecraft/world/gen/chunk/ChunkGenerator;Ljava/util/Random;Lnet/minecraft/util/math/BlockBox;Lnet/minecraft/util/math/ChunkPos;Lnet/minecraft/util/math/BlockPos;)V",
             at = @At(value = "INVOKE_ASSIGN", target = "Lnet/minecraft/world/StructureWorldAccess;getTopY(Lnet/minecraft/world/Heightmap$Type;II)I", ordinal = 0),
             cancellable = true, locals = LocalCapture.CAPTURE_FAILHARD
     )
     private void structurefloaters_disableHeightmapSnap(StructureWorldAccess world, StructureAccessor structureAccessor,
                                                    ChunkGenerator chunkGenerator, Random random, BlockBox boundingBox,
-                                                   ChunkPos chunkPos, BlockPos pos, CallbackInfoReturnable<Boolean> cir,
+                                                   ChunkPos chunkPos, BlockPos pos, CallbackInfo ci,
                                                    int i, int j)
     {
         if(!StructureFloaters.STRUCTURES_TO_IGNORE.contains(SHIPWRECK_ID) &&
                 StructureFloaters.SF_CONFIG.removeStructuresOffIslands &&
                 chunkGenerator.getSeaLevel() <= chunkGenerator.getMinimumY() &&
-                j <= world.getBottomY())
+                j <= world.getBottomY() + 1)
         {
-            cir.setReturnValue(false);
+            ci.cancel();
         }
     }
 
@@ -73,7 +74,7 @@ public abstract class ShipwreckGeneratorPieceMixin extends SimpleStructurePiece 
      * @reason Prevent structures from being placed at world bottom if disallowed in config
      */
     @ModifyVariable(
-            method = "generate(Lnet/minecraft/world/StructureWorldAccess;Lnet/minecraft/world/gen/StructureAccessor;Lnet/minecraft/world/gen/chunk/ChunkGenerator;Ljava/util/Random;Lnet/minecraft/util/math/BlockBox;Lnet/minecraft/util/math/ChunkPos;Lnet/minecraft/util/math/BlockPos;)Z",
+            method = "generate(Lnet/minecraft/world/StructureWorldAccess;Lnet/minecraft/world/gen/StructureAccessor;Lnet/minecraft/world/gen/chunk/ChunkGenerator;Ljava/util/Random;Lnet/minecraft/util/math/BlockBox;Lnet/minecraft/util/math/ChunkPos;Lnet/minecraft/util/math/BlockPos;)V",
             at = @At(value = "INVOKE_ASSIGN", target = "Lnet/minecraft/world/StructureWorldAccess;getTopY(Lnet/minecraft/world/Heightmap$Type;II)I", ordinal = 0),
             ordinal = 1
     )
@@ -84,7 +85,7 @@ public abstract class ShipwreckGeneratorPieceMixin extends SimpleStructurePiece 
         if(!StructureFloaters.STRUCTURES_TO_IGNORE.contains(SHIPWRECK_ID) &&
                 chunkGenerator.getSeaLevel() <= chunkGenerator.getMinimumY() &&
                 !(StructureFloaters.SF_CONFIG.removeStructuresOffIslands &&
-                world.getTopY(type, this.pos.getX(), this.pos.getZ()) <= chunkGenerator.getMinimumY()) &&
+                world.getTopY(type, this.pos.getX(), this.pos.getZ()) <= chunkGenerator.getMinimumY() + 1) &&
                 heightmapY < StructureFloaters.SF_CONFIG.snapStructureToHeight)
         {
             return StructureFloaters.SF_CONFIG.snapStructureToHeight;
@@ -97,22 +98,22 @@ public abstract class ShipwreckGeneratorPieceMixin extends SimpleStructurePiece 
      * @reason Prevent structures from being placed at world bottom if disallowed in config
      */
     @Inject(
-            method = "generate(Lnet/minecraft/world/StructureWorldAccess;Lnet/minecraft/world/gen/StructureAccessor;Lnet/minecraft/world/gen/chunk/ChunkGenerator;Ljava/util/Random;Lnet/minecraft/util/math/BlockBox;Lnet/minecraft/util/math/ChunkPos;Lnet/minecraft/util/math/BlockPos;)Z",
+            method = "generate(Lnet/minecraft/world/StructureWorldAccess;Lnet/minecraft/world/gen/StructureAccessor;Lnet/minecraft/world/gen/chunk/ChunkGenerator;Ljava/util/Random;Lnet/minecraft/util/math/BlockBox;Lnet/minecraft/util/math/ChunkPos;Lnet/minecraft/util/math/BlockPos;)V",
             at = @At(value = "INVOKE_ASSIGN", target = "Lnet/minecraft/world/StructureWorldAccess;getTopY(Lnet/minecraft/world/Heightmap$Type;II)I", ordinal = 1, shift = At.Shift.AFTER),
             cancellable = true, locals = LocalCapture.CAPTURE_FAILHARD
     )
     private void structurefloaters_disableHeightmapSnap2(StructureWorldAccess world, StructureAccessor structureAccessor,
                                                     ChunkGenerator chunkGenerator, Random random, BlockBox boundingBox,
-                                                    ChunkPos chunkPos, BlockPos pos, CallbackInfoReturnable<Boolean> cir,
+                                                    ChunkPos chunkPos, BlockPos pos, CallbackInfo ci,
                                                     int i, int j, Vec3i vec3i, Heightmap.Type type, int k, BlockPos blockPos,
                                                     Iterator<BlockPos> var14, BlockPos blockPos2, int l)
     {
         if(!StructureFloaters.STRUCTURES_TO_IGNORE.contains(SHIPWRECK_ID) &&
                 StructureFloaters.SF_CONFIG.removeStructuresOffIslands &&
                 chunkGenerator.getSeaLevel() <= chunkGenerator.getMinimumY() &&
-                l <= world.getBottomY())
+                l <= world.getBottomY() + 1)
         {
-            cir.setReturnValue(false);
+            ci.cancel();
         }
     }
 
@@ -122,7 +123,7 @@ public abstract class ShipwreckGeneratorPieceMixin extends SimpleStructurePiece 
      * @reason Prevent structures from being placed at world bottom if disallowed in config
      */
     @ModifyVariable(
-            method = "generate(Lnet/minecraft/world/StructureWorldAccess;Lnet/minecraft/world/gen/StructureAccessor;Lnet/minecraft/world/gen/chunk/ChunkGenerator;Ljava/util/Random;Lnet/minecraft/util/math/BlockBox;Lnet/minecraft/util/math/ChunkPos;Lnet/minecraft/util/math/BlockPos;)Z",
+            method = "generate(Lnet/minecraft/world/StructureWorldAccess;Lnet/minecraft/world/gen/StructureAccessor;Lnet/minecraft/world/gen/chunk/ChunkGenerator;Ljava/util/Random;Lnet/minecraft/util/math/BlockBox;Lnet/minecraft/util/math/ChunkPos;Lnet/minecraft/util/math/BlockPos;)V",
             at = @At(value = "INVOKE_ASSIGN", target = "Lnet/minecraft/world/StructureWorldAccess;getTopY(Lnet/minecraft/world/Heightmap$Type;II)I", ordinal = 1),
             ordinal = 3
     )
@@ -133,7 +134,7 @@ public abstract class ShipwreckGeneratorPieceMixin extends SimpleStructurePiece 
         if(!StructureFloaters.STRUCTURES_TO_IGNORE.contains(SHIPWRECK_ID) &&
                 chunkGenerator.getSeaLevel() <= chunkGenerator.getMinimumY() &&
                 !(StructureFloaters.SF_CONFIG.removeStructuresOffIslands &&
-                        world.getTopY(type, this.pos.getX(), this.pos.getZ()) <= chunkGenerator.getMinimumY()) &&
+                        world.getTopY(type, this.pos.getX(), this.pos.getZ()) <= chunkGenerator.getMinimumY() + 1) &&
                 heightmapY < StructureFloaters.SF_CONFIG.snapStructureToHeight)
         {
             return StructureFloaters.SF_CONFIG.snapStructureToHeight;